backend/requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Document processing
pdfplumber==0.11.0
python-docx==1.1.2

# Vector store + embeddings
chromadb==0.5.3
sentence-transformers==3.0.1

# LLM / Ollama
httpx==0.27.0
httpx-sse==0.4.0

# Database
aiosqlite==0.20.0
sqlalchemy[asyncio]==2.0.30

# Structured output
pydantic==2.7.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
